customer/customer_app/serializers.py
from rest_framework import serializers
from customer_app.models import Customer

class CustomerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Customer
        fields = ['name', 'username', 'email']