customer/requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
pillow==10.2.0
psycopg2==2.9.9
Pygments==2.17.2
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
